Add files via upload
This is the result after fixing the resolution of images used in training. 512x512 images are fed into the system, and only the output is scaled (before, the input images were being scaled, which lead to images being too small for tent detection). It still struggles with tent detection, but performs significantly better, tending to at least generally recognize regions that contain tents. It seems to often confuse cars for tents as well.